+++ resources/views/cams/index.blade.php
@@ -35,8 +35,8 @@
                                 </a>
                             </p>
                             <p class="control">
-                                {!! Form::open(['method' => 'DELETE','route' => ['cam.destroy', $cam->id]]) !!}
-                                <button type="submit" class="button is-danger is-small">
+                                {!! Form::open(['method' => 'DELETE','route' => ['cam.destroy', $cam->id],'onsubmit' => 'return confirm("Do you really want to delete this cctv?");']) !!}
+                                <button type="submit" class="button is-danger is-small" >
                                 <span class="icon is-small">
                                     <i class="fa fa-trash-o"></i>
                                 </span>
