+++ resources/views/divisions/index.blade.php
@@ -25,7 +25,7 @@
                                 </a>
                             </p>
                             <p class="control">
-                                {!! Form::open(['method' => 'DELETE','route' => ['division.destroy', $division->id]]) !!}
+                                {!! Form::open(['method' => 'DELETE','route' => ['division.destroy', $division->id],'onsubmit' => 'return confirm("Do you really want to delete this division?");']) !!}
                                 <button type="submit" class="button is-danger is-small">
                                     <span class="icon is-small">
                                         <i class="fa fa-trash-o"></i>
